fixed updatemoodupdate and updatemooddelete utils to use the correct date (sent from controllers) instead of today and changed naming too

+++ memora-backend/src/Utils/UpdateDailyMoodDelete.js
@@ -19,16 +19,18 @@ function calculateMood(score){
     return "excited";
 }
 
-export const updateDailyMoodDelete = async (score, userID, timezone) => {
+export const updateDailyMoodDelete = async (score, userID, timezone, entryDate) => {
     
     const tz = timezone || "UTC";
     
-    const today = dayjs().tz(tz).startOf("day").utc().toDate();
+    const date = dayjs(entryDate).tz(tz).startOf("day").utc().toDate();
 
     const dailyMood = await DailyMood.findOne({
         user: userID,
-        date: today,
-    })
+        date,
+    });
+
+    if (!dailyMood) return;
 
     if (dailyMood.entries === 1) {
         await dailyMood.deleteOne();
